requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
python-dotenv==1.0.0

